refactor(assignments): extrai useLotteryRun e helpers puros de participante
As derivações transversais do sorteio (participantIds, filters, configKey,
prévia/seed, eligibleCount, buildParams, handlePreview/handleRandomize) e os
estados loading/previewing saem do componente para useLotteryRun, que
desestrutura o LotteryParamsState mantendo os arrays de deps dos memos
idênticos. isParticipant/weightValue/capValue viram funções puras em
lottery-participant-values.ts, compartilhadas entre a seção Participantes e
participantSettings (fonte única da coerção). LotteryDialog: 953→657 linhas.

Refs #357
